Phase_1/requirement.txt
fastapi
uvicorn
transformers
torch
torchvision
decord
Pillow
numpy
accelerate
qwen-vl-utils
python-multipart